Fix struct definition long->int
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@811 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/src/DES.cs
@@ -369,7 +369,7 @@ namespace DirectShowLib.DES
     [StructLayout(LayoutKind.Sequential, Pack=4)]
     public class SCompFmt0
     {
-        public long nFormatId;
+        public int nFormatId;
         public AMMediaType MediaType;
     }
